Fix issues with non-ASCII filenames on exporting

+++ src/cpp/core/http/Util.cpp
@@ -230,7 +230,7 @@ std::string urlEncode(const std::string& in, bool queryStringSpaces)
          std::ostringstream ostr ;
          ostr << "%" ;
          ostr << std::setw(2) << std::setfill('0') << std::hex << std::uppercase
-              << (int)ch ;
+              << (int)(uint8_t)ch ;
          std::string charAsHex = ostr.str();
          encodedURL += charAsHex;
       }
